Renaming/moving misc components
